Migrate class to private inner class of only client
